Extract desk context menu state resolver
